+++ test/IntegrationTests/AspNetTests.cs
@@ -15,10 +15,15 @@
 // </copyright>
 
 #if NETFRAMEWORK
+using System;
 using System.IO;
+using System.Linq;
 using System.Net.Http;
 using System.Threading.Tasks;
+using FluentAssertions;
+using FluentAssertions.Execution;
 using IntegrationTests.Helpers;
+using Opentelemetry.Proto.Common.V1;
 using Xunit;
 using Xunit.Abstractions;
 
@@ -46,8 +51,12 @@ public class AspNetTests : TestHelper
         // on the firewall.
         using var agent = new MockZipkinCollector(Output, host: "*");
         using var fwPort = FirewallHelper.OpenWinPort(agent.Port, Output);
+        var testSettings = new TestSettings
+        {
+            TracesSettings = new TracesSettings { Port = agent.Port }
+        };
         var webPort = TcpPortProvider.GetOpenPort();
-        using var container = await StartContainerAsync(agent.Port, webPort);
+        using var container = await StartContainerAsync(testSettings, webPort);
 
         var client = new HttpClient();
 
@@ -62,6 +71,45 @@ public class AspNetTests : TestHelper
         var spans = agent.WaitForSpans(1);
 
         Assert.True(spans.Count >= 1, $"Expecting at least 1 span, only received {spans.Count}");
+    }
+
+    [Fact]
+    [Trait("Category", "EndToEnd")]
+    [Trait("Containers", "Windows")]
+    public async Task SubmitMetrics()
+    {
+        const int expectedMetricRequests = 1;
+
+        // Using "*" as host requires Administrator. This is needed to make the mock collector endpoint
+        // accessible to the Windows docker container where the test application is executed by binding
+        // the endpoint to all network interfaces. In order to do that it is necessary to open the port
+        // on the firewall.
+        using var collector = new MockMetricsCollector(Output, host: "*");
+        using var fwPort = FirewallHelper.OpenWinPort(collector.Port, Output);
+        var testSettings = new TestSettings
+        {
+            MetricsSettings = new MetricsSettings { Port = collector.Port },
+        };
+        var webPort = TcpPortProvider.GetOpenPort();
+        using var container = await StartContainerAsync(testSettings, webPort);
+
+        var client = new HttpClient();
+
+        var response = await client.GetAsync($"http://localhost:{webPort}");
+        var content = await response.Content.ReadAsStringAsync();
+
+        Output.WriteLine("Sample response:");
+        Output.WriteLine(content);
+
+        var metricRequests = collector.WaitForMetrics(expectedMetricRequests, TimeSpan.FromSeconds(5));
+
+        using (new AssertionScope())
+        {
+            metricRequests.Count.Should().BeGreaterThanOrEqualTo(expectedMetricRequests);
+            var resourceMetrics = metricRequests.SelectMany(r => r.ResourceMetrics).Where(s => s.ScopeMetrics.Count > 0).FirstOrDefault();
+            var aspnetMetrics = resourceMetrics.ScopeMetrics.Should().ContainSingle(x => x.Scope.Name == "OpenTelemetry.Instrumentation.AspNet").Which.Metrics;
+            aspnetMetrics.Should().ContainSingle(x => x.Name == "http.server.duration");
+        }
     }
 }
 #endif